+++ routes/vehicle.js
@@ -7,21 +7,23 @@ router.get('/', (req, res) => {
 
 router.get('/:vehicle_id', (req, res) => {
   db.query(`SELECT * FROM cars
-    WHERE idCar="${req.params.vehicle_id}"
-    AND active>0;`, (err, result) => {
-    if (!result.length)  return res.render('page_not_found', { user: req.user });
+    WHERE idCar LIKE ?
+    AND active NOT LIKE 0;`, req.params.vehicle_id, (err, result) => {
+    if (!result.length)
+      return res.render('page_not_found', { user: req.user });
     else {
       db.query(`SELECT idCar, photo FROM cars
-        WHERE idCar NOT LIKE "${req.params.vehicle_id}"
-        AND active>0
+        WHERE idCar NOT LIKE ?
+        AND active NOT LIKE 0
         ORDER BY RAND()
-        LIMIT 10;`, (err1, result1) => {
+        LIMIT 10;`, req.params.vehicle_id, (err1, result1) => {
         if (req.user) {
           db.query(`SELECT * FROM reservations
-            WHERE idCar="${req.params.vehicle_id}"
-            AND idUser="${req.user.idUser}"
+            WHERE idCar LIKE ?
+            AND idUser LIKE ?
             AND DATE(dateOut)>=CURDATE()
-            AND active>0;`, (err2, result2) => {
+            AND active NOT LIKE 0;`, [req.params.vehicle_id,
+              req.user.idUser], (err2, result2) => {
             return res.render('vehicle', {
               car: result[0],
               carsArr: result1,
@@ -30,10 +32,10 @@ router.get('/:vehicle_id', (req, res) => {
             });
           });
         } else return res.render('vehicle', {
-          car: result[0],
-          carsArr: result1,
-          resv: null,
-          user: req.user
+            car: result[0],
+            carsArr: result1,
+            resv: null,
+            user: req.user
         });
       });
     }
@@ -41,6 +43,8 @@ router.get('/:vehicle_id', (req, res) => {
 });
 
 router.post('/add/:car_id', (req, res) => {
+  if (!req.user) return res.redirect('/');
+  
   const resv = {
     active: 1,
     idUser: req.user.idUser,
@@ -50,26 +54,31 @@ router.post('/add/:car_id', (req, res) => {
     fullPrice: req.body.numbers
   };
 
-  db.query('INSERT into reservations SET ?', resv, (err) => {
+  db.query(`INSERT into reservations SET ?`, resv, (err) => {
     if (err) return res.send({ code: 400, failed: __print('add', err)});
     return res.redirect('/library?res=add');
   });
 });
 
 router.post('/edit/:res_id/', (req, res) => {
+  if (!req.user) return res.redirect('/');
+
   db.query(`UPDATE reservations
-    SET fullPrice="${req.body.numbers}",
-      dateIn="${req.body.dateIn}",
-      dateOut="${req.body.dateOut}"
-    WHERE idReservation="${req.params.res_id}";`, (err) => {
+    SET fullPrice=?,
+      dateIn=?,
+      dateOut=?
+    WHERE idReservation LIKE ?;`, [req.body.numbers,
+      req.body.dateIn, req.body.dateOut, req.params.res_id], (err) => {
     if (err) return res.send({ code: 400, failed: __print('edit', err)});
     return res.redirect('/library?res=edit');
   });
 });
 
 router.post('/rem/:res_id/', (req, res) => {
+  if (!req.user) return res.redirect('/');
+
   db.query(`DELETE FROM reservations
-    WHERE idReservation="${req.params.res_id}";`, (err) => {
+    WHERE idReservation LIKE ?;`, req.params.res_id, (err) => {
     if (err) return res.send({ code: 400, failed: __print('delete', err)});
     return res.redirect('/library?res=del');
   });
